#38 New achievement groups: guess people from n different countries, guess x people, you haven't guessed previously.

+++ app/src/main/java/rjm/romek/facegame/achievement/ConsecutiveDaysPlayingAchievementUpdater.java
@@ -12,6 +12,6 @@ public class ConsecutiveDaysPlayingAchievementUpdater extends SqlPoweredUpdater 
 
     @Override
     protected Long transform(QuestionService change) {
-        return (long) change.countConsecutiveDaysPlaying(numberOfDaysToUnlock);
+        return change.countConsecutiveDaysPlaying(numberOfDaysToUnlock);
     }
 }
